Move "constants" into src/internal - move constants into "src/internal/constants" - add small note on what it is
