+++ src/index.ts
@@ -1,7 +1,8 @@
+import { DecoratorKeys } from './internal/constants';
+
 /*
  copy-paste from mongodb package (should be same as IndexOptions from 'mongodb')
  */
-import { DecoratorKeys } from './constants';
 
 export interface IndexOptions<T> {
   /**
