+++ src/internal/constants.ts
@@ -1,3 +1,6 @@
+/**
+ * Collection of Reflect Types for easy maintenance
+ */
 export enum DecoratorKeys {
   Prop = 'design:type',
   ModelOptions = 'typegoose:options',
