+++ src/typegoose.ts
@@ -10,12 +10,12 @@ if (!Object.fromEntries) {
 
 import { deprecate } from 'util';
 import * as defaultClasses from './defaultClasses';
+import { DecoratorKeys } from './internal/constants';
 import { buildSchemas, constructors, models } from './internal/data';
 import { _buildSchema } from './internal/schema';
+import { assignMetadata } from './internal/utils';
 import { IModelOptions } from './optionsProp';
 import { DocumentType, NoParamConstructor, Ref, ReturnModelType } from './types';
-import { assignMetadata } from './internal/utils';
-import { DecoratorKeys } from './constants';
 
 /* exports */
 export * from './method';
